PM/requirements2.txt
Django==1.10.5
MySQL-python==1.2.5
PAM==0.4.2
Pillow==2.3.0
PyMySQL==0.7.9
Twisted-Core==13.2.0
Twisted-Web==13.2.0
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
debtagshw==0.1
defer==1.0.6
dirspec==13.10
django-appconf==1.0.2
django-bootstrap3==8.1.0
django-cities-light==3.3.0
django-countries==4.0
django-imagekit==3.3
django-phonenumber-field==1.1.0
django-social-auth==0.7.28
django-sortedm2m==1.3.3
djangorestframework==3.6.2
duplicity==0.6.23
html5lib==0.999
httplib2==0.8
lockfile==0.8
lxml==3.3.3
mysql-connector-python==1.1.6
oauth2==1.9.0.post1
oauthlib==0.6.1
oneconf==0.3.7.14.04.1
pexpect==3.1
phonenumbers==8.4.2
pilkit==1.1.13
piston-mini-client==0.7.5
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pygobject==3.12.0
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5ubuntu2
python-debian==0.1.21-nmu2ubuntu2
python-openid==2.2.5
pyxdg==0.25
reportlab==3.0
requests==2.2.1
sessioninstaller==0.0.0
six==1.5.2
software-center-aptd-plugins==0.0.0
sorl-thumbnail==12.3
system-service==0.1.6
unity-lens-photos==1.0
urllib3==1.7.1
virtualenv==15.1.0
wheel==0.24.0
wsgiref==0.1.2
xdiagnose==3.6.3build2
zope.interface==4.0.5
